requirements.txt
openai
lxml
python-docx
EbookLib
beautifulsoup4
pytest
